No need to spread a module around.
