--- lib/alchemist/exceptions.rb
@@ -1,5 +1,0 @@
-module Alchemist
-
-  class YamlSyntaxError < StandardError; end
-
-end
